backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
langchain
langchain-community
langchain-google-genai
google-generativeai
faiss-cpu
sentence-transformers
pydantic
pypdf
